requirements.txt
amqp==2.3.2
APScheduler==3.5.3
Babel==2.5.1
billiard==3.5.0.3
celery==4.2.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
idna==2.6
kombu==4.2.2
pydash==4.7.4
PyJWT==1.7.1
pymongo==3.7.2
python-dotenv==0.10
pytz==2017.3
requests==2.21.0
six==1.11.0
tzlocal==1.5.1
urllib3==1.24.1
vine==1.1.4
